OMIGOD NetworkManager and dhcpd were BOTH getting a DHCP lease!
This is so fucked.  On picky captive portal networks, that'd cause me
to get 2 IP addresses, and the routing table would have whatever the
most recent DHCP lease gave as "src" address.  This would cause
problems since that was also the "secondary" address in `ip addr`.
Still unsure if that's it, but... geez.